First static version for large scale test.
TODO: Try improve the accuracy.

+++ src/main/java/edu/uq/dke/mapupdate/mapinference/BiagioniKDE2012.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 public class BiagioniKDE2012 {
     private KDEMapInference mapInference;
 
-    public BiagioniKDE2012(double cellSize, int blur) {
+    public BiagioniKDE2012(int cellSize, int blur) {
         this.mapInference = new KDEMapInference(cellSize, blur);
     }
